updating sign up and log in endopoints

+++ back/internal/core/models/user/user.go
@@ -3,7 +3,7 @@ package user
 import "time"
 
 type User struct {
-	ID        uint64   `gorm:"primaryKey"`
+	ID        uint32   `gorm:"primaryKey"`
 	Name      UserName `gorm:"not null"`
 	Email     Email    `gorm:"unique"`
 	Password  Password `gorm:"not null"`
